share params
