+++ web/app/components/plugins/marketplace/atoms.ts
@@ -1,7 +1,7 @@
 import type { PluginsSort } from './types'
 import { atom, useAtom, useAtomValue, useSetAtom } from 'jotai'
-import { useMarketplaceCategory, useMarketplaceSearchQuery, useMarketplaceTags } from '@/hooks/use-query-params'
 import { DEFAULT_SORT, PLUGIN_CATEGORY_WITH_COLLECTIONS } from './constants'
+import { useMarketplaceCategory, useMarketplaceSearchQuery, useMarketplaceTags } from './hooks'
 
 const marketplaceSortAtom = atom<PluginsSort>(DEFAULT_SORT)
