+++ web/app/components/plugins/marketplace/constants.ts
@@ -1,3 +1,5 @@
+import { parseAsArrayOf, parseAsString } from 'nuqs/server'
+
 import { PluginCategoryEnum } from '../types'
 
 export const DEFAULT_SORT = {
@@ -24,3 +26,9 @@ export const PLUGIN_CATEGORY_WITH_COLLECTIONS = new Set(
     PLUGIN_TYPE_SEARCH_MAP.tool,
   ],
 )
+
+export const marketplaceSearchParams = {
+  q: parseAsString.withDefault('').withOptions({ history: 'replace' }),
+  category: parseAsString.withDefault('all').withOptions({ history: 'replace', clearOnDefault: false }),
+  tags: parseAsArrayOf(parseAsString).withDefault([]).withOptions({ history: 'replace' }),
+}
