+++ web/app/components/plugins/marketplace/plugin-type-switch.tsx
@@ -9,10 +9,10 @@ import {
   RiSpeakAiLine,
 } from '@remixicon/react'
 import { Trigger as TriggerIcon } from '@/app/components/base/icons/src/vender/plugin'
-import { useMarketplaceCategory } from '@/hooks/use-query-params'
 import { cn } from '@/utils/classnames'
 import { useSetSearchMode } from './atoms'
 import { PLUGIN_CATEGORY_WITH_COLLECTIONS, PLUGIN_TYPE_SEARCH_MAP } from './constants'
+import { useMarketplaceCategory } from './hooks'
 
 type PluginTypeSwitchProps = {
   className?: string
